# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mausam 2.0 - Personalized Weather Intelligence",
  "description": "Mausam 2.0 - a Personalized Weather Intelligence Engine on official IMD/MoES data. Personalized homepage cards for the 'Mausam' app (SIH 2026 PS 26076).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/users/{user_id}/profile": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Get Profile",
    "operationId": "get_profile_api_v1_users__user_id__profile_get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProfileSyncOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "users"
    ],
    "summary": "Put Profile",
    "operationId": "put_profile_api_v1_users__user_id__profile_put",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileSync"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProfileSyncOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "users"
    ],
    "summary": "Delete Profile",
    "description": "Permanently remove this user's cloud backup (right-to-erasure).",
    "operationId": "delete_profile_api_v1_users__user_id__profile_delete",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Delete Profile Api V1 Users  User Id  Profile Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{user_id}/push-token": {
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Register Push Token",
    "description": "Register the device's Expo/FCM push token for server->device alerts.",
    "operationId": "register_push_token_api_v1_users__user_id__push_token_post",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PushTokenIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Register Push Token Api V1 Users  User Id  Push Token Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "users"
    ],
    "summary": "Unregister Push Token",
    "operationId": "unregister_push_token_api_v1_users__user_id__push_token_delete",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PushTokenIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Unregister Push Token Api V1 Users  User Id  Push Token Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{user_id}/notifications": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "List Notifications",
    "description": "The durable alert inbox for this user (server-side archive).",
    "operationId": "list_notifications_api_v1_users__user_id__notifications_get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationListOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{user_id}/notifications/{notification_id}/ack": {
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Ack Notification",
    "description": "Mark a single archive entry as read (clears its unread badge).",
    "operationId": "ack_notification_api_v1_users__user_id__notifications__notification_id__ack_post",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Notification Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Ack Notification Api V1 Users  User Id  Notifications  Notification Id  Ack Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/admin/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_admin_users_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Get User",
    "operationId": "get_user_api_admin_users__user_id__get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user_api_admin_users__user_id__put",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Delete User",
    "operationId": "delete_user_api_admin_users__user_id__delete",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AppNotification": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "alert_id": {
      "type": "string",
      "title": "Alert Id"
     },
     "severity": {
      "$ref": "#/components/schemas/WarningSeverity"
     },
     "event_type": {
      "type": "string",
      "title": "Event Type"
     },
     "headline": {
      "type": "string",
      "title": "Headline"
     },
     "body": {
      "type": "string",
      "title": "Body"
     },
     "region": {
      "type": "string",
      "title": "Region"
     },
     "read": {
      "type": "boolean",
      "title": "Read",
      "default": false
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "alert_id",
     "severity",
     "event_type",
     "headline",
     "body",
     "region",
     "created_at"
    ],
    "title": "AppNotification"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NotificationListOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "unread": {
      "type": "integer",
      "title": "Unread",
      "default": 0
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/AppNotification"
      },
      "type": "array",
      "title": "Items",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "user_id"
    ],
    "title": "NotificationListOut"
   },
...
   "ProfileSync": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "default": ""
     },
     "name_hi": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name Hi"
     },
     "personas": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Personas",
      "default": []
     },
     "conditions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Conditions",
      "default": []
     },
     "activities": {
      "items": {
       "$ref": "#/components/schemas/SyncActivity"
      },
      "type": "array",
      "title": "Activities",
      "default": []
     },
     "locations": {
      "items": {
       "$ref": "#/components/schemas/SyncLocation"
      },
      "type": "array",
      "title": "Locations",
      "default": []
     },
     "city": {
      "type": "string",
      "title": "City",
      "default": "pune"
     },
     "language": {
      "type": "string",
      "title": "Language",
      "default": "en"
     },
     "behavior_bias": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Behavior Bias",
      "default": {}
     }
    },
    "type": "object",
    "required": [
     "id"
    ],
    "title": "ProfileSync",
    "description": "The mobile UserProfile payload used to move a profile between devices.\n\nThe client stores this locally (AsyncStorage) and can PUT it to the server,\nthen GET it back on another device to restore the same experience."
   },
   "ProfileSyncOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "profile": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ProfileSync"
       },
       {
        "type": "null"
       }
      ]
     },
     "updated_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "user_id"
    ],
    "title": "ProfileSyncOut"
   },
...
   "PushTokenIn": {
    "properties": {
     "token": {
      "type": "string",
      "maxLength": 512,
      "minLength": 8,
      "title": "Token"
     },
     "platform": {
      "type": "string",
      "pattern": "^(android|ios)$",
      "title": "Platform",
      "default": "android"
     }
    },
    "type": "object",
    "required": [
     "token"
    ],
    "title": "PushTokenIn"
   },
...
   "SyncActivity": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type"
     },
     "label": {
      "type": "string",
      "title": "Label",
      "default": ""
     },
     "label_hi": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label Hi"
     },
     "time": {
      "type": "string",
      "title": "Time",
      "default": "09:00"
     }
    },
    "type": "object",
    "required": [
     "type"
    ],
    "title": "SyncActivity",
    "description": "A scheduled activity on the mobile profile."
   },
   "SyncLocation": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type",
      "default": "home"
     },
     "label": {
      "type": "string",
      "title": "Label",
      "default": ""
     }
    },
    "type": "object",
    "title": "SyncLocation",
    "description": "A saved place (home/work/school/farm) on the mobile profile."
   },
   "UserUpdate": {
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "personas": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Personas"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "health": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Health"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WarningSeverity": {
    "type": "string",
    "enum": [
     "green",
     "yellow",
     "orange",
     "red"
    ],
    "title": "WarningSeverity"
   },
...
  }
 }
}